+++ src/features/profile/hooks/useBookmarkRemove.ts
@@ -5,6 +5,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import type { Bookmark } from '@/shared/types/bookmark';
 import { removeBookmark } from '../api/bookmarksApi';
+import { BOOKMARKS_QUERY_KEY } from './useScrapedJobs';
 import { useCurrentUser } from '@/shared/hooks/useCurrentUser';
 
 export function useBookmarkRemove() {
@@ -16,11 +17,12 @@ export function useBookmarkRemove() {
     mutationFn: (postingUrl: string) => removeBookmark(postingUrl),
 
     onMutate: async (postingUrl) => {
-      await queryClient.cancelQueries({ queryKey: ['bookmarks'] });
+      await queryClient.cancelQueries({ queryKey: BOOKMARKS_QUERY_KEY });
 
-      const previous = queryClient.getQueryData<Bookmark[]>(['bookmarks']);
+      const previous =
+        queryClient.getQueryData<Bookmark[]>(BOOKMARKS_QUERY_KEY);
 
-      queryClient.setQueryData<Bookmark[]>(['bookmarks'], (old = []) =>
+      queryClient.setQueryData<Bookmark[]>(BOOKMARKS_QUERY_KEY, (old = []) =>
         old.filter((b) => b.postingUrl !== postingUrl),
       );
 
@@ -29,7 +31,7 @@ export function useBookmarkRemove() {
 
     onError: (_err, _url, ctx) => {
       if (ctx?.previous) {
-        queryClient.setQueryData(['bookmarks'], ctx.previous);
+        queryClient.setQueryData(BOOKMARKS_QUERY_KEY, ctx.previous);
       }
     },
   });
